refactor: 提取权限验证辅助函数到utils.go
- 新建backend/internal/service/utils.go
- 移动getOrgIDFromContext和verifyOrgPermission到utils.go
- 清理digital_assistant_service.go中的空行
- 将接口验证移到文件顶部
